reorganize code

+++ app/src/main/java/com/jdmobile/inflightsalesapp/domain/model/Product.kt
@@ -2,14 +2,15 @@ package com.jdmobile.inflightsalesapp.domain.model
 
 import com.jdmobile.inflightsalesapp.ui.screens.product.model.ProductFilter
 import kotlinx.serialization.Serializable
+import java.math.BigDecimal
 
 data class Product(
     val id: ProductId,
     val name: String,
     val stock: Int,
-    val priceUSD: Double,
-    val priceEUR: Double,
-    val priceGBP: Double,
+    val priceUSD: BigDecimal,
+    val priceEUR: BigDecimal,
+    val priceGBP: BigDecimal,
     val imageUrl: String,
     val category: ProductFilter
 )
